Add Subversion support.
This is using an older version of the subversion libraries. I think the better approach is to use the ‘svn’ command line tool because these libraries are a heavy dependency for only getting Subversion status and we need to keep them updated, as Subversion periodically upgrade the repository format in a non-backwards compatible way.